[WIP][#66] All code refactored, created new namespace: app\helpers\RememberUserInfo; Achievements RememberHelper added; fixed couple of inaccuracies in view/site/login.php, view/layouts/main.php
